+++ src/kaya/collection.rb
