+++ src/kaya/factory.rb
