+++ src/logger.rb
