+++ src/timetable_collection.rb
