+++ src/timetable_factory.rb
