+++ src/timetable_fetcher.rb
