+++ src/timetable_item.rb
