+++ src/timetable_place.rb
